refactor(#581): extract ExerciseDetailModal.vue from WorkoutTracker
Extract the exercise detail bottom sheet (sets list, PR history, graph,
swipe-to-dismiss, focus trap) into a standalone ExerciseDetailModal
component. WorkoutTracker passes exerciseId and receives events for
actions that mutate parent state (log set, edit exercise, edit set,
delete set).

Reduces WorkoutTracker from 3389 to 3160 lines (-229). The new
component owns its own tab state, swipe gesture, focus trap, warmup
filtering, PR computation, and set grouping.
